feat: move getProductRoute
